Create aedc_for_anchda
This code is written in R programming language and it is used to analyze data from the Australian Early Development Census (AEDC). The AEDC is a nationwide survey conducted every three years that assesses the developmental status of children in their first year of full-time school. The aim of this code is to calculate the prevalence of developmental vulnerabilities in different domains (physical health and wellbeing, social competence, emotional maturity, language and cognitive skills, and communication skills and general knowledge) across different Local Government Areas (LGAs) in Australia.

The code first loads the necessary libraries including readxl, stringr, dplyr, purrr, and readr. Then, it sets the working directory to the folder where the input data is stored, and imports the data from a csv file using the read.csv function. The mutate function is then used to create new variables for each developmental domain, based on the values in the original variable for that domain. For example, a new variable DOT_PHW is created based on the PHYSCategory variable, and takes a value of 1 if the PHYSCategory is 3 or 4, and 0 otherwise.

The code then calculates the prevalence of developmental vulnerabilities in each LGA for each developmental domain, based on the new variables created earlier. For example, for the physical health and wellbeing domain, the code uses the summarize function to calculate the number and proportion of children with a vulnerability in each LGA, as well as the number and proportion of children with a developmental delay or disorder (DAR), and the number and proportion of children with a developmental vulnerability only (DOT). This is repeated for each developmental domain.

The results are stored in separate data frames LGA_df_PHW, LGA_df_SC, LGA_df_EM, LGA_df_LCS, and LGA_df_CSGK, which contain the prevalence estimates for each LGA, year, and gender. These data frames are saved in separate csv files in the output folder specified by path_out.